Move the FileServer extension into ReconnectCore
